requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.1.31
click==8.1.7
ecdsa==0.19.1
Faker==37.1.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
outcome==1.3.0.post0
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
PySocks==1.7.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9.1
selenium==4.31.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
starlette==0.41.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.32.1
websocket-client==1.8.0
wsproto==1.2.0
